requirements.txt
asgiref==3.8.1
cloudinary==1.41.0
dj-database-url==2.3.0
Django==5.1.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.1
whitenoise==6.7.0
